count all tools in changed repo to get chunk count
currently the chunk count depends only on the number of changed tools.
but the tool test uses all tools in the changed repos. therefore the
computed number of chunks is to low.

for the linting step we still need the changed tools (for the .shed.yml
test).

for CI (i.e. when $COMMIT_RANGE is empty) the two calls to ci_find_tools
yield identical results and we do not need to run ci_find_tools again